refactor: State와 Running, Finished 사이에 공통 부분을 AbstractState로 묶어줌
