+++ src/main/java/blackjack/domain/state/Finished.java
@@ -1,17 +1,12 @@
 package blackjack.domain.state;
 
-import blackjack.domain.Score;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
-import java.util.Objects;
 
-public abstract class Finished implements State {
-
-    private final Cards cards;
+public abstract class Finished extends AbstractState {
 
     protected Finished(Cards cards) {
-        Objects.requireNonNull(cards, "[ERROR] 카드패는 null일 수 없습니다.");
-        this.cards = cards;
+        super(cards);
     }
 
     @Override
@@ -27,15 +22,5 @@ public abstract class Finished implements State {
     @Override
     public boolean isFinished() {
         return true;
-    }
-
-    @Override
-    public Cards getCards() {
-        return cards;
-    }
-
-    @Override
-    public Score calculateScore() {
-        return cards.calculateScore();
     }
 }
